+++ bitmart/lib/cloud_consts.py
@@ -114,6 +114,12 @@ API_CONTRACT_CANCEL_TRAIL_ORDER_URL = "/contract/private/cancel-trail-order"
 
 API_CONTRACT_TRANSFER_CONTRACT_LIST_URL = "/account/v1/transfer-contract-list"
 API_CONTRACT_TRANSFER_CONTRACT_URL = "/account/v1/transfer-contract"
+API_CONTRACT_CANCEL_ALL_AFTER_URL = "/contract/private/cancel-all-after"
+API_CONTRACT_GET_POSITION_MODE_URL = "/contract/private/get-position-mode"
+API_CONTRACT_SET_POSITION_MODE_URL = "/contract/private/set-position-mode"
+API_CONTRACT_POSITION_V2_URL = "/contract/private/position-v2"
+API_CONTRACT_LEVERAGE_BRACKET_URL = "/contract/public/leverage-bracket"
+API_CONTRACT_MARKET_TRADE_URL = "/contract/public/market-trade"
 
 
 # websocket