+++ examples/futures/market/get_market_trade.py
@@ -1,27 +1,16 @@
 import logging
-from examples.config import API_KEY, SECRET_KEY, MEMO
-import time
 
 from bitmart.api_contract import APIContract
 from bitmart.lib.cloud_exceptions import APIException
 from bitmart.lib.cloud_utils import config_logging
 
-
 config_logging(logging, logging.DEBUG)
 logger = logging.getLogger(__name__)
 
-futuresAPI = APIContract(api_key=API_KEY,
-                         secret_key=SECRET_KEY,
-                         memo=MEMO,
-                         logger=logger)
+contractAPI = APIContract(logger=logger)
 
 try:
-    response = futuresAPI.post_transfer(
-        currency="USDT",
-        amount="100",
-        type="spot_to_contract",
-        recv_window=6000,
-    )[0]
+    response = contractAPI.get_market_trade(contract_symbol='BTCUSDT', limit=2)[0]
     logger.info(response)
 except APIException as error:
     logger.error(
@@ -29,4 +18,3 @@ except APIException as error:
             error.status_code, error.response
         )
     )
-