+++ examples/futures/trade/cancel_trail_order.py
@@ -1,17 +1,19 @@
 import logging
+from examples.config import API_KEY, SECRET_KEY, MEMO
 import time
 
 from bitmart.api_contract import APIContract
 from bitmart.lib.cloud_exceptions import APIException
 from bitmart.lib.cloud_utils import config_logging
 
+
 config_logging(logging, logging.DEBUG)
 logger = logging.getLogger(__name__)
 
 
-futuresAPI = APIContract(api_key="Your_Api_Key",
-                         secret_key="Your_Secret_Key",
-                         memo="Your_Memo",
+futuresAPI = APIContract(api_key=API_KEY,
+                         secret_key=SECRET_KEY,
+                         memo=MEMO,
                          logger=logger)
 
 try: