+++ examples/futures/trade/get_current_position_risk.py
@@ -1,16 +1,18 @@
 import logging
+from examples.config import API_KEY, SECRET_KEY, MEMO
 import time
 
 from bitmart.api_contract import APIContract
 from bitmart.lib.cloud_exceptions import APIException
 from bitmart.lib.cloud_utils import config_logging
 
+
 config_logging(logging, logging.DEBUG)
 logger = logging.getLogger(__name__)
 
-futuresAPI = APIContract(api_key="Your_Api_Key",
-                         secret_key="Your_Secret_Key",
-                         memo="Your_Memo",
+futuresAPI = APIContract(api_key=API_KEY,
+                         secret_key=SECRET_KEY,
+                         memo=MEMO,
                          logger=logger)
 
 try:
@@ -26,6 +28,19 @@ except APIException as error:
 try:
     response = futuresAPI.get_position_risk(
         contract_symbol='BTCUSDT',
+    )[0]
+    logger.info(response)
+except APIException as error:
+    logger.error(
+        "Found error. status: {}, error message: {}".format(
+            error.status_code, error.response
+        )
+    )
+
+try:
+    response = futuresAPI.get_position_risk(
+        contract_symbol='BTCUSDT',
+        account='futures'
     )[0]
     logger.info(response)
 except APIException as error: