+++ examples/spot/margin/get_repayment_record_isolated.py
@@ -1,15 +1,17 @@
 import logging
+from examples.config import API_KEY, SECRET_KEY, MEMO
 
 from bitmart.api_margin_loan import APIMarginLoan
 from bitmart.lib.cloud_exceptions import APIException
 from bitmart.lib.cloud_utils import config_logging
 
+
 config_logging(logging, logging.DEBUG)
 logger = logging.getLogger(__name__)
 
-marginLoanAPI = APIMarginLoan(api_key="Your_Api_Key",
-                              secret_key="Your_Secret_Key",
-                              memo="Your_Memo",
+marginLoanAPI = APIMarginLoan(api_key=API_KEY,
+                              secret_key=SECRET_KEY,
+                              memo=MEMO,
                               logger=logger)
 
 try: