+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 Changelog
 =========================
 
+### v2.5.0 Release
+#### New Features
+- New
+    - `/contract/private/cancel-all-after` Applicable for canceling all contract orders timed(SIGNED)
+    - `/contract/private/get-position-mode` Get Position Mode (KEYED)
+    - `/contract/private/set-position-mode` Set Position Mode (SIGNED)
+    - `/contract/private/position-v2` Get Current Position V2 (KEYED)
+    - `/contract/public/leverage-bracket` Get Current Leverage Risk Limit
+    - `/contract/public/market-trade` Query the latest trade data
+- Update
+    - `/contract/private/order` Get Order Detail (KEYED): Add new request field **account**
+    - `/contract/private/order-history` Get Order History (KEYED): Add new request field **account** and **order_id** and **client_order_id**
+    - `/contract/private/position` Get Current Position (KEYED): Add new request field **account**
+    - `/contract/private/position-risk` Get Current Position Risk Details(KEYED): Add new request field **account**
+    - `/contract/private/trades` Get Order Trade (KEYED): Add new request field **account**, request field **symbol** changed to optional
+    - `/contract/private/transaction-history` Get Transaction History (KEYED): Add new request field **account**
+
+
+---
 
 ### v2.4.2 Release
 #### Improvements